Remove Content-Type header from image fetches
That header is for describing the content being sent with the request (of which there is none). 'Accept' is what tells the server what types of content it should send, but since the fetch is for a specific file, it's unlikely the server will be able to send a different image format, anyway. At best it could redirect, but that seems unlikely, and at worst, it could refuse to serve the file.