Clean out no longer relevant items
Remove references to websites that are no longer in operation, as well as any scripts that were specific to them. This greatly reduces the set of scripts that need to be adjusted to make work with later browser versions, and reduces permissions cruft in the manifest. There's probably a lot of CDN hosts that can be removed, too, but that can wait until run-time permissions is implemented.