Make work with current Firefox, Chrome versions
This required a little refactoring to shunt all uses of fetch() or XMLHttpRequest through a couple common functions that will defer them to the background script unless they are requests for same origin as the page that is running.

Also, now there is a background script on Firefox and a service worker on Chrome.

Switched to manifest v3 so Chrome doesn't stop working in a month or 2 when Chrome disables v2. Also for Chrome only, had to add declarativeNetRequest permission to get referrer set in fetch()es made by service worker.

These changes also make mangakomi.io work, which was the last remaining broken site that is still in operation.

Also, rename the extension slightly, so it doesn't conflict with the original.