Add packaging script
Add a script for packaging up just the files needed for the Chrome extension, and just those for the Firefox extension. The repository will work as-is on either Chrome or Firefox, but that will result in a warning about unknown property in the manifest. Also, Firefox doesn't need a permission that Chrome does.