Convert site-specific scripts to a common template
Move all the not-website-specific code from the website-specific scripts into a generic template script, since most of the scripts were doing the exact same things. Now the website-specific scripts define functions to implement the website-specific bits.

Improve the handling of chapter list pages that have hundreds, or even thousands, of chapters by deferring the fetch of the image URL list until either the individual download button is pressed or the chapter is selected for batch download. Also, split up adding the buttons to 100 chapters at a time, to give the browser time to render the page.

As a side effect of this change, some chapter list pages that did not have batch download buttons now have them, because that was easier than coding support for not having it. Also, the websites that have multiple server options for the download buttons now support batch download from the same options, rather than forcing batch download to use the first server.