Use the manganelo content servers for manganelo
The manganato content servers seem to work just fine, but these are the ones that will be used when reading manga on mangenelo.com. While chasing down what turned out to be an unrelated problem, I thought this might be necessary now. It's not, but it still makes more sense to do it this way.